requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
anthropic==0.34.2
pdfplumber==0.11.4
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2
python-multipart==0.0.9
python-dotenv==1.0.1
openpyxl==3.1.5
openai==1.51.0
bcrypt==4.2.0
PyJWT==2.9.0
